requirements.txt
python-dotenv==1.0.1
selenium>=4.34.2